+++ libexplain/buffer/fildes_to_pathname.c
@@ -1,6 +1,6 @@
 /*
  * libexplain - Explain errno values returned by libc functions
- * Copyright (C) 2008, 2009 Peter Miller
+ * Copyright (C) 2008-2010 Peter Miller
  * Written by Peter Miller <pmiller@opensource.org.au>
  *
  * This program is free software; you can redistribute it and/or modify
@@ -17,8 +17,9 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include <libexplain/ac/limits.h> /* for PATH_MAX on Solaris */
 #include <libexplain/ac/stdio.h>
-#include <libexplain/ac/sys/param.h>
+#include <libexplain/ac/sys/param.h> /* for PATH_MAX except Solaris */
 #include <libexplain/ac/sys/socket.h>
 #include <libexplain/ac/sys/stat.h>
 #include <libexplain/ac/unistd.h>
@@ -75,7 +76,7 @@ explain_buffer_fildes_to_path(explain_string_buffer_t *sb, int fildes)
 
     obj.inherited.n_callback = n_callback;
     obj.sb = sb;
-    snprintf(options, sizeof(options), "-p %d -d %d", getpid(), fildes);
+    snprintf(options, sizeof(options), "-p %ld -d %d", (long)getpid(), fildes);
     explain_lsof(options, &obj.inherited);
 #endif
 }
