+++ libexplain/message/write.c
@@ -1,6 +1,6 @@
 /*
  * libexplain - Explain errno values returned by libc functions
- * Copyright (C) 2008, 2009 Peter Miller
+ * Copyright (C) 2008-2010 Peter Miller
  * Written by Peter Miller <pmiller@opensource.org.au>
  *
  * This program is free software; you can redistribute it and/or modify
@@ -18,13 +18,14 @@
  */
 
 #include <libexplain/ac/errno.h>
+#include <libexplain/ac/unistd.h>
 
 #include <libexplain/write.h>
 
 
 void
 explain_message_write(char *message, int message_size, int fildes,
-    const void *data, long data_size)
+    const void *data, size_t data_size)
 {
     explain_message_errno_write
     (
