+++ libexplain/write.h
@@ -1,16 +1,16 @@
 /*
  * libexplain - Explain errno values returned by libc functions
- * Copyright (C) 2008, 2009 Peter Miller
+ * Copyright (C) 2010 Peter Miller
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU Lesser General Public License as published by
  * the Free Software Foundation; either version 3 of the License, or (at
  * your option) any later version.
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser
+ * General Public License for more details.
  *
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
@@ -22,9 +22,15 @@
 /**
   * @file
   * @brief explain write(2) errors
+  *
+  * These functions may be used to obtain explanations for errors returned
+  * by the <i>write</i>(2) system call.
   */
 
 #include <libexplain/warn_unused_result.h>
+#include <libexplain/large_file_support.h>
+
+#include <unistd.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -33,14 +39,8 @@ extern "C" {
 /**
   * The explain_write_or_die function is used to call the <i>write</i>(2)
   * system call. On failure an explanation will be printed to stderr,
-  * obtained from the explain_write(3) function, and then the process
-  * terminates by calling exit(EXIT_FAILURE).
-  *
-  * This function is intended to be used in a fashion similar to the
-  * following example:
-  * @code
-  * ssize_t result = explain_write_or_die(fildes, data, data_size);
-  * @endcode
+  * obtained from the <i>#explain_write</i>(3) function, and then the
+  * process terminates by calling <tt>exit(EXIT_FAILURE)</tt>.
   *
   * @param fildes
   *     The fildes, exactly as to be passed to the <i>write</i>(2) system
@@ -52,26 +52,23 @@ extern "C" {
   *     The data_size, exactly as to be passed to the <i>write</i>(2)
   *     system call.
   * @returns
-  *     This function only returns on success. On failure, prints an
-  *     explanation and exits, it does not return.
+  *     This function only returns on success, see <i>write</i>(2) for more
+  *     information. On failure, prints an explanation and exits, it does
+  *     not return.
+  *
+  * @par Example:
+  * This function is intended to be used in a fashion similar to the
+  * following example:
+  * @code
+  * ssize_t result = explain_write_or_die(fildes, data, data_size);
+  * @endcode
   */
-long explain_write_or_die(int fildes, const void *data, long data_size);
+ssize_t explain_write_or_die(int fildes, const void *data, size_t data_size);
 
 /**
   * The explain_write_on_error function is used to call the <i>write</i>(2)
   * system call. On failure an explanation will be printed to stderr,
-  * obtained from the explain_write(3) function.
-  *
-  * This function is intended to be used in a fashion similar to the
-  * following example:
-  * @code
-  * ssize_t result = explain_write_on_error(fildes, data, data_size);
-  * if (result < 0)
-  * {
-  *     ...cope with error
-  *     ...no need to print error message
-  * }
-  * @endcode
+  * obtained from the <i>#explain_write</i>(3) function.
   *
   * @param fildes
   *     The fildes, exactly as to be passed to the <i>write</i>(2) system
@@ -84,8 +81,20 @@ long explain_write_or_die(int fildes, const void *data, long data_size);
   *     system call.
   * @returns
   *     The value returned by the wrapped <i>write</i>(2) system call.
+  *
+  * @par Example:
+  * This function is intended to be used in a fashion similar to the
+  * following example:
+  * @code
+  * ssize_t result = explain_write_on_error(fildes, data, data_size);
+  * if (result < 0)
+  * {
+  *     ...cope with error
+  *     ...no need to print error message
+  * }
+  * @endcode
   */
-int explain_write_on_error(int fildes, const void *data, long data_size)
+ssize_t explain_write_on_error(int fildes, const void *data, size_t data_size)
                                                   LIBEXPLAIN_WARN_UNUSED_RESULT;
 
 /**
@@ -96,20 +105,6 @@ int explain_write_on_error(int fildes, const void *data, long data_size)
   *
   * The errno global variable will be used to obtain the error value to be
   * decoded.
-  *
-  * This function is intended to be used in a fashion similar to the
-  * following example:
-  * @code
-  * ssize_t resuilt = write(fildes, data, data_size);
-  * if (result < 0)
-  * {
-  *     fprintf(stderr, "%s\n", explain_write(fildes, data, data_size));
-  *     exit(EXIT_FAILURE);
-  * }
-  * @endcode
-  *
-  * The above code example is available pre-packaged as the
-  * #explain_write_or_die function.
   *
   * @param fildes
   *     The original fildes, exactly as passed to the <i>write</i>(2)
@@ -130,8 +125,23 @@ int explain_write_on_error(int fildes, const void *data, long data_size)
   *     This function is <b>not</b> thread safe, because it shares a return
   *     buffer across all threads, and many other functions in this
   *     library.
+  *
+  * @par Example:
+  * This function is intended to be used in a fashion similar to the
+  * following example:
+  * @code
+  * ssize_t result = write(fildes, data, data_size);
+  * if (result < 0)
+  * {
+  *     fprintf(stderr, "%s\n", explain_write(fildes, data, data_size));
+  *     exit(EXIT_FAILURE);
+  * }
+  * @endcode
+  * @par
+  * The above code example is available pre-packaged as the
+  * #explain_write_or_die function.
   */
-const char *explain_write(int fildes, const void *data, long data_size)
+const char *explain_write(int fildes, const void *data, size_t data_size)
                                                   LIBEXPLAIN_WARN_UNUSED_RESULT;
 
 /**
@@ -140,22 +150,6 @@ const char *explain_write(int fildes, const void *data, long data_size)
   * message will contain is the value of <tt>strerror(errnum)</tt>, but
   * usually it will do much better, and indicate the underlying cause in
   * more detail.
-  *
-  * This function is intended to be used in a fashion similar to the
-  * following example:
-  * @code
-  * ssize_t result = write(fildes, data, data_size);
-  * if (result < 0)
-  * {
-  *     int err = errno;
-  *     fprintf(stderr, "%s\n", explain_errno_write(err, fildes, data,
-  *         data_size));
-  *     exit(EXIT_FAILURE);
-  * }
-  * @endcode
-  *
-  * The above code example is available pre-packaged as the
-  * #explain_write_or_die function.
   *
   * @param errnum
   *     The error value to be decoded, usually obtained from the errno
@@ -182,9 +176,26 @@ const char *explain_write(int fildes, const void *data, long data_size)
   *     This function is <b>not</b> thread safe, because it shares a return
   *     buffer across all threads, and many other functions in this
   *     library.
+  *
+  * @par Example:
+  * This function is intended to be used in a fashion similar to the
+  * following example:
+  * @code
+  * ssize_t result = write(fildes, data, data_size);
+  * if (result < 0)
+  * {
+  *     int err = errno;
+  *     fprintf(stderr, "%s\n", explain_errno_write(err, fildes, data,
+  *         data_size));
+  *     exit(EXIT_FAILURE);
+  * }
+  * @endcode
+  * @par
+  * The above code example is available pre-packaged as the
+  * #explain_write_or_die function.
   */
-const char *explain_errno_write(int errnum, int fildes, const void *data,
-    long data_size)
+const char *explain_errno_write(int errnum, int fildes, const void *data, size_t
+    data_size)
                                                   LIBEXPLAIN_WARN_UNUSED_RESULT;
 
 /**
@@ -196,23 +207,6 @@ const char *explain_errno_write(int errnum, int fildes, const void *data,
   *
   * The errno global variable will be used to obtain the error value to be
   * decoded.
-  *
-  * This function is intended to be used in a fashion similar to the
-  * following example:
-  * @code
-  * ssize_t result = write(fildes, data, data_size);
-  * if (result < 0)
-  * {
-  *     char message[3000];
-  *     explain_message_write(message, sizeof(message), fildes, data,
-  *         data_size);
-  *     fprintf(stderr, "%s\n", message);
-  *     exit(EXIT_FAILURE);
-  * }
-  * @endcode
-  *
-  * The above code example is available pre-packaged as the
-  * #explain_write_or_die function.
   *
   * @param message
   *     The location in which to store the returned message. If a suitable
@@ -229,9 +223,27 @@ const char *explain_errno_write(int errnum, int fildes, const void *data,
   * @param data_size
   *     The original data_size, exactly as passed to the <i>write</i>(2)
   *     system call.
+  *
+  * @par Example:
+  * This function is intended to be used in a fashion similar to the
+  * following example:
+  * @code
+  * ssize_t result = write(fildes, data, data_size);
+  * if (result < 0)
+  * {
+  *     char message[3000];
+  *     explain_message_write(message, sizeof(message), fildes, data,
+  *         data_size);
+  *     fprintf(stderr, "%s\n", message);
+  *     exit(EXIT_FAILURE);
+  * }
+  * @endcode
+  * @par
+  * The above code example is available pre-packaged as the
+  * #explain_write_or_die function.
   */
-void explain_message_write(char *message, int message_size, int fildes,
-    const void *data, long data_size);
+void explain_message_write(char *message, int message_size, int fildes, const
+    void *data, size_t data_size);
 
 /**
   * The explain_message_errno_write function is used to obtain an
@@ -240,6 +252,29 @@ void explain_message_write(char *message, int message_size, int fildes,
   * <tt>strerror(errnum)</tt>, but usually it will do much better, and
   * indicate the underlying cause in more detail.
   *
+  * @param message
+  *     The location in which to store the returned message. If a suitable
+  *     message return buffer is supplied, this function is thread safe.
+  * @param message_size
+  *     The size in bytes of the location in which to store the returned
+  *     message.
+  * @param errnum
+  *     The error value to be decoded, usually obtained from the errno
+  *     global variable just before this function is called. This is
+  *     necessary if you need to call <b>any</b> code between the system
+  *     call to be explained and this function, because many libc functions
+  *     will alter the value of errno.
+  * @param fildes
+  *     The original fildes, exactly as passed to the <i>write</i>(2)
+  *     system call.
+  * @param data
+  *     The original data, exactly as passed to the <i>write</i>(2) system
+  *     call.
+  * @param data_size
+  *     The original data_size, exactly as passed to the <i>write</i>(2)
+  *     system call.
+  *
+  * @par Example:
   * This function is intended to be used in a fashion similar to the
   * following example:
   * @code
@@ -254,34 +289,12 @@ void explain_message_write(char *message, int message_size, int fildes,
   *     exit(EXIT_FAILURE);
   * }
   * @endcode
-  *
+  * @par
   * The above code example is available pre-packaged as the
   * #explain_write_or_die function.
-  *
-  * @param message
-  *     The location in which to store the returned message. If a suitable
-  *     message return buffer is supplied, this function is thread safe.
-  * @param message_size
-  *     The size in bytes of the location in which to store the returned
-  *     message.
-  * @param errnum
-  *     The error value to be decoded, usually obtained from the errno
-  *     global variable just before this function is called. This is
-  *     necessary if you need to call <b>any</b> code between the system
-  *     call to be explained and this function, because many libc functions
-  *     will alter the value of errno.
-  * @param fildes
-  *     The original fildes, exactly as passed to the <i>write</i>(2)
-  *     system call.
-  * @param data
-  *     The original data, exactly as passed to the <i>write</i>(2) system
-  *     call.
-  * @param data_size
-  *     The original data_size, exactly as passed to the <i>write</i>(2)
-  *     system call.
   */
 void explain_message_errno_write(char *message, int message_size, int errnum,
-    int fildes, const void *data, long data_size);
+    int fildes, const void *data, size_t data_size);
 
 #ifdef __cplusplus
 }
